Update README.md
This application gives the user ability to select their very own password length. This can be useful for those who create online accounts and would like to automatically create a password of the standard eight characters. If users prefer a password longer than 8 characters, they may do so. 

With this simple program, say goodbye to compromised information and data when creating passwords.